refactor(chat): split live runtime tests
